Fix uint64 mobile build issue
