+++ zcncore/smartcontract_executor.go
@@ -3,6 +3,8 @@
 
 package zcncore
 
+import "github.com/0chain/errors"
+
 type SmartContractExecutor interface {
 	// ExecuteSmartContract implements wrapper for smart contract function
 	ExecuteSmartContract(address, methodName string, input interface{}, val uint64) error
@@ -28,4 +30,27 @@ func (ta *TransactionWithAuth) ExecuteSmartContract(address, methodName string, 
 		ta.submitTxn()
 	}()
 	return nil
+}
+
+// NewTransaction allocation new generic transaction object for any operation
+func NewTransaction(cb TransactionCallback, txnFee uint64, nonce int64) (TransactionScheme, error) {
+	err := CheckConfig()
+	if err != nil {
+		return nil, err
+	}
+	if _config.isSplitWallet {
+		if _config.authUrl == "" {
+			return nil, errors.New("", "auth url not set")
+		}
+		Logger.Info("New transaction interface with auth")
+		return newTransactionWithAuth(cb, txnFee, nonce)
+	}
+	Logger.Info("New transaction interface")
+	t, err := newTransaction(cb, txnFee, nonce)
+	return t, err
+}
+
+// ConvertToValue converts ZCN tokens to value
+func ConvertToValue(token float64) uint64 {
+	return uint64(token * float64(TOKEN_UNIT))
 }
