+++ zcncore/smartcontract_executor_mobile.go
@@ -3,7 +3,13 @@
 
 package zcncore
 
-import "encoding/json"
+import (
+	"encoding/json"
+	"fmt"
+	"strconv"
+
+	"github.com/0chain/errors"
+)
 
 type SmartContractExecutor interface {
 	// ExecuteSmartContract implements wrapper for smart contract function
@@ -30,4 +36,36 @@ func (ta *TransactionWithAuth) ExecuteSmartContract(address, methodName string, 
 		ta.submitTxn()
 	}()
 	return nil
+}
+
+// NewTransaction allocation new generic transaction object for any operation
+func NewTransaction(cb TransactionCallback, txnFee string, nonce int64) (TransactionScheme, error) {
+	fee, err := strconv.ParseUint(txnFee, 10, 64)
+	if err != nil {
+		return nil, fmt.Errorf("invalid transaction fee value: %v", txnFee)
+	}
+
+	if fee/uint64(TOKEN_UNIT) == 0 {
+		return nil, fmt.Errorf("transaction fee must be multiple value of 1e10")
+	}
+
+	err = CheckConfig()
+	if err != nil {
+		return nil, err
+	}
+	if _config.isSplitWallet {
+		if _config.authUrl == "" {
+			return nil, errors.New("", "auth url not set")
+		}
+		Logger.Info("New transaction interface with auth")
+		return newTransactionWithAuth(cb, fee, nonce)
+	}
+	Logger.Info("New transaction interface")
+	t, err := newTransaction(cb, fee, nonce)
+	return t, err
+}
+
+// ConvertToValue converts ZCN tokens to value
+func ConvertToValue(token float64) string {
+	return strconv.FormatUint(uint64(token*float64(TOKEN_UNIT)), 10)
 }
