+++ zcncore/transaction.go
@@ -379,24 +379,6 @@ func newTransaction(cb TransactionCallback, txnFee uint64, nonce int64) (*Transa
 	return t, nil
 }
 
-// NewTransaction allocation new generic transaction object for any operation
-func NewTransaction(cb TransactionCallback, txnFee uint64, nonce int64) (TransactionScheme, error) {
-	err := CheckConfig()
-	if err != nil {
-		return nil, err
-	}
-	if _config.isSplitWallet {
-		if _config.authUrl == "" {
-			return nil, errors.New("", "auth url not set")
-		}
-		Logger.Info("New transaction interface with auth")
-		return newTransactionWithAuth(cb, txnFee, nonce)
-	}
-	Logger.Info("New transaction interface")
-	t, err := newTransaction(cb, txnFee, nonce)
-	return t, err
-}
-
 func (t *Transaction) SetTransactionCallback(cb TransactionCallback) error {
 	if t.txnStatus != StatusUnknown {
 		return errors.New("", "transaction already exists. cannot set transaction hash.")
